fix(test) Webpack-plugin tests
