+++ package.json
@@ -73,7 +73,7 @@
     "typescript": "5.7.3",
     "typescript-eslint": "8.24.0",
     "webpack": "5.97.1",
-    "webpack4": "npm:webpack@^4.42.1"
+    "webpack4": "npm:webpack@^4.47.0"
   },
   "dependencies": {
     "@bundle-stats/plugin-webpack-filter": "4.18.2",
