+++ test/webpack-plugin-webpack-v4.test.js
@@ -4,25 +4,12 @@ jest.mock('webpack', () => require('webpack4')); // eslint-disable-line
 const webpack = require('webpack4');
 const MemoryFS = require('memory-fs');
 const fetch = require('isomorphic-fetch');
-const { merge } = require('lodash');
 
-const pckg = require('../package.json');
 const webpackStats = require('./__snapshots__/webpack-4-stats.json');
 const appConfig = require('./webpack/webpack.config');
-
-const ENV_DEFAULT = {
-  CI: 'true',
-  RELATIVE_CI_ENDPOINT: 'http://localhost/save',
-  RELATIVE_CI_KEY: '123',
-  CIRCLECI: 'true',
-  CIRCLE_SHA1: 'abcd1234',
-  CIRCLE_BRANCH: 'master',
-  CIRCLE_PROJECT_USERNAME: 'organization',
-  CIRCLE_PROJECT_REPONAME: 'project',
-  CIRCLE_BUILD_URL: '#',
-  CIRCLE_BUILD_NUM: 123,
-  CIRCLE_PR_NUMBER: 10,
-};
+const {
+  ENV_DEFAULT, clearCustomEnv, getMockRequest, setCustomEnv,
+} = require('./utils');
 
 const MOCK_RESULT = {
   res: {
@@ -35,48 +22,6 @@ const MOCK_RESULT = {
       txt: 'Hello world!',
     },
   },
-};
-
-const getMockRequest = (customPayload) => ({
-  method: 'POST',
-  headers: {
-    'Content-Type': 'application/json; charset=utf-8',
-  },
-  body: JSON.stringify(merge(
-    {
-      key: '123',
-      project: 'organization/project',
-      service: 'circleci',
-      job: {
-        commit: 'abcd1234',
-        branch: 'master',
-        prNumber: '10',
-        buildNumber: '123',
-        buildUrl: '#',
-      },
-      rawData: {
-        webpack: {
-          stats: {},
-        },
-      },
-      agentVersion: pckg.version,
-    },
-    customPayload,
-  )),
-});
-
-const setCustomEnv = (customEnv = {}) => {
-  const envVars = { ...ENV_DEFAULT, ...customEnv };
-
-  Object.entries(envVars).forEach(([key, value]) => {
-    process.env[key] = value;
-  });
-};
-
-const clearCustomEnv = () => {
-  Object.keys(ENV_DEFAULT).forEach((key) => {
-    process.env[key] = undefined;
-  });
 };
 
 describe('webpack-plugin', () => {
