+++ test/webpack-plugin-webpack-v5.test.js
@@ -3,25 +3,12 @@ jest.mock('isomorphic-fetch');
 const webpack = require('webpack');
 const MemoryFS = require('memory-fs');
 const fetch = require('isomorphic-fetch');
-const { merge } = require('lodash');
 
-const pckg = require('../package.json');
 const webpack5Stats = require('./__snapshots__/webpack-5-stats.json');
+const {
+  ENV_DEFAULT, clearCustomEnv, getMockRequest, setCustomEnv,
+} = require('./utils');
 const appConfig = require('./webpack/webpack.config');
-
-const ENV_DEFAULT = {
-  CI: 'true',
-  RELATIVE_CI_ENDPOINT: 'http://localhost/save',
-  RELATIVE_CI_KEY: '123',
-  CIRCLECI: 'true',
-  CIRCLE_SHA1: 'abcd1234',
-  CIRCLE_BRANCH: 'master',
-  CIRCLE_PROJECT_USERNAME: 'organization',
-  CIRCLE_PROJECT_REPONAME: 'project',
-  CIRCLE_BUILD_URL: '#',
-  CIRCLE_BUILD_NUM: 123,
-  CIRCLE_PR_NUMBER: 10,
-};
 
 const MOCK_RESULT = {
   res: {
@@ -34,48 +21,6 @@ const MOCK_RESULT = {
       txt: 'Hello world!',
     },
   },
-};
-
-const getMockRequest = (customPayload) => ({
-  method: 'POST',
-  headers: {
-    'Content-Type': 'application/json; charset=utf-8',
-  },
-  body: JSON.stringify(merge(
-    {
-      key: '123',
-      project: 'organization/project',
-      service: 'circleci',
-      job: {
-        commit: 'abcd1234',
-        branch: 'master',
-        prNumber: '10',
-        buildNumber: '123',
-        buildUrl: '#',
-      },
-      rawData: {
-        webpack: {
-          stats: {},
-        },
-      },
-      agentVersion: pckg.version,
-    },
-    customPayload,
-  )),
-});
-
-const setCustomEnv = (customEnv = {}) => {
-  const envVars = { ...ENV_DEFAULT, ...customEnv };
-
-  Object.entries(envVars).forEach(([key, value]) => {
-    process.env[key] = value;
-  });
-};
-
-const clearCustomEnv = () => {
-  Object.keys(ENV_DEFAULT).forEach((key) => {
-    process.env[key] = undefined;
-  });
 };
 
 describe('webpack-plugin', () => {
